requirements.txt
altgraph==0.17.2
asgiref==3.4.1
astroid==2.5.2
certifi==2021.10.8
colorama==0.4.4
Django==3.2.9
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
future==0.18.2
importlib-metadata==4.8.2
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pefile==2021.9.3
pylint==2.7.4
pymongo==3.12.1
pytz==2021.3
pywin32-ctypes==0.2.0
sqlparse==0.2.4
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.10.0.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
